move measurements panel
